preliminary sh(4) support
currently only the sh4 variant is supported, as sh3 requires
a nofpu subarch, and i'm not sure how to best check for that.
the correct ld suffix would be -nofpu, i.e. ld-musl-sh-nofpu.so.1.

to build correctly, the gcc triple must be set to sh4-linux-musl,
otherwise the build system will try to build a sh1/sh2 toolchain.
sh1/2 is currently broken, the toolchain will default to soft-imask
atomic model, which raises this error:
"error: cannot use atomic model soft-imask in user mode"
(besides it has no mmu, so it is not usable for linux anyway).